+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Change Log
 
+## v1.2.0 :rocket:
+
+- Added support for controller partials
+    * Files with a leading underscore are considered partials and will be ignored
+    * Should be used to share functionality between controllers
+- Asset directory is now recursively copied from the source directory to the output directory
+- Added support for sass stylesheets
+    * Files can be `.scss` or `.sass` files
+    * Files with a leading underscore are considered partials and are ignored
+    * `@import` resolves files in the stylesheet dir or in `node_modules`
+    * stylesheets must be placed in the `stylesheets` subdirectory in your assets dir
+- Output directory is now cleaned between builds
+- Removed Herobrine
+
 ## v1.1.2 :rocket:
 
 - Fixed bugs preventing rendering from working